Add files via upload
In the past, if we were to start writing a Python app, our first order of business was to install a Python runtime onto our machine. But, that creates a situation where the environment on our machine needs to be perfect for our app to run as expected, and also needs to match our production environment.

With Docker, we can just grab a portable Python runtime as an image, no installation necessary. Then, our build can include the base Python image right alongside our app code, ensuring that our app, its dependencies, and the runtime, all travel together.

To get into project we first understood what containers are, the difference between the vm and containers. Then the next step was to create a container.which started with creating a dockerfile.Dockerfile is a file that you create which in turn produces a Docker image when you build it then the act of running a Docker image creates a Docker container.